Keep held-out LoRA tests off auxiliary perf probe
